[border-agent] move `EphemeralKeyManager` to its own files
Moves `EphemeralKeyManager` class and its implementation from
`border_agent.hpp` and `border_agent.cpp` to their own separate files
`border_agent_ephemeral_key.hpp/cpp`.

This is a structural change to improve code organization and does not
introduce any functional changes.
